backend/requirement.txt
fastapi
uvicorn[standard]
python-dotenv
groq
pydantic
requests
SQLAlchemy

